Add runtime option to disable pool trackin per-pool
